Solve day 15
It's satisfying to have to think for once… not so much to implement the
binary space search, but hey it was worth it for that moment of victory.

Side note: I must have left the brute force method running for nearly
half a day and it didn't complete. I guess that's the point, but it's
cool to have seen good optimization working well.