Separate out each day
Due to my optimistic aproach of expecting each day to parse input into a
struct with a trait and Advent of Code having a rather fast-paced
development cycle, I ended up abandoning structure at some point in
favour of just doing things all in the main function.

This only moves the code, without really changing any of the structure.
It improves things slightly, but is mostly an aesthetic change.